Shopping cart.

+++ src/main/java/com/sagafitmi/ecommerce/service/ProductService.java
@@ -3,6 +3,7 @@ package com.sagafitmi.ecommerce.service;
 import java.util.List;
 
 import com.sagafitmi.ecommerce.dto.ProductDTO;
+import java.math.BigDecimal;
 
 public interface ProductService {
     // Define service methods here
@@ -10,6 +11,6 @@ public interface ProductService {
     ProductDTO getProductById(Long id);
     ProductDTO createProduct(ProductDTO productDTO);
     ProductDTO updateProduct(Long id, ProductDTO productDTO);
-    ProductDTO updateProductPrice(Long id, Double price);
+    ProductDTO updateProductPrice(Long id, BigDecimal price);
     void deleteProduct(Long id);
 }
